requirements.txt
# Web框架和服务器
fastapi==0.103.2
uvicorn==0.22.0
pydantic==2.0.3
python-multipart==0.0.6

# 图像处理
pillow==11.2.1

# HTTP请求和网络
requests==2.32.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.4.0

# 实用工具
python-dotenv==1.0.0
click==8.1.7
h11==0.14.0

# 部署
gunicorn==21.2.0 